+++ newrelic/newrelic.go
@@ -7,31 +7,32 @@ import (
 
 	log "github.com/sirupsen/logrus"
 
-	"github.com/newrelic/newrelic-client-go/pkg/accounts"
-	"github.com/newrelic/newrelic-client-go/pkg/alerts"
-	"github.com/newrelic/newrelic-client-go/pkg/apiaccess"
-	"github.com/newrelic/newrelic-client-go/pkg/apm"
-	"github.com/newrelic/newrelic-client-go/pkg/cloud"
-	"github.com/newrelic/newrelic-client-go/pkg/config"
-	"github.com/newrelic/newrelic-client-go/pkg/dashboards"
-	"github.com/newrelic/newrelic-client-go/pkg/edge"
-	"github.com/newrelic/newrelic-client-go/pkg/entities"
-	"github.com/newrelic/newrelic-client-go/pkg/events"
-	"github.com/newrelic/newrelic-client-go/pkg/eventstometrics"
-	"github.com/newrelic/newrelic-client-go/pkg/installevents"
-	"github.com/newrelic/newrelic-client-go/pkg/logging"
-	"github.com/newrelic/newrelic-client-go/pkg/logs"
-	"github.com/newrelic/newrelic-client-go/pkg/nerdgraph"
-	"github.com/newrelic/newrelic-client-go/pkg/nerdstorage"
-	"github.com/newrelic/newrelic-client-go/pkg/notifications"
-	"github.com/newrelic/newrelic-client-go/pkg/nrdb"
-	"github.com/newrelic/newrelic-client-go/pkg/nrqldroprules"
-	"github.com/newrelic/newrelic-client-go/pkg/plugins"
-	"github.com/newrelic/newrelic-client-go/pkg/region"
-	"github.com/newrelic/newrelic-client-go/pkg/servicelevel"
-	"github.com/newrelic/newrelic-client-go/pkg/synthetics"
-	"github.com/newrelic/newrelic-client-go/pkg/workflows"
-	"github.com/newrelic/newrelic-client-go/pkg/workloads"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/accounts"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/alerts"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/apiaccess"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/apm"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/changetracking"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/cloud"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/config"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/dashboards"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/edge"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/entities"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/events"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/eventstometrics"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/installevents"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/logging"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/logs"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/nerdgraph"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/nerdstorage"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/notifications"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/nrdb"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/nrqldroprules"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/plugins"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/region"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/servicelevel"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/synthetics"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/workflows"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/workloads"
 )
 
 // NewRelic is a collection of New Relic APIs.
@@ -40,6 +41,7 @@ type NewRelic struct {
 	Alerts          alerts.Alerts
 	APIAccess       apiaccess.APIAccess
 	APM             apm.APM
+	ChangeTracking  changetracking.Changetracking
 	Cloud           cloud.Cloud
 	Dashboards      dashboards.Dashboards
 	Edge            edge.Edge
@@ -90,6 +92,7 @@ func New(opts ...ConfigOption) (*NewRelic, error) {
 		Alerts:          alerts.New(cfg),
 		APIAccess:       apiaccess.New(cfg),
 		APM:             apm.New(cfg),
+		ChangeTracking:  changetracking.New(cfg),
 		Cloud:           cloud.New(cfg),
 		Dashboards:      dashboards.New(cfg),
 		Edge:            edge.New(cfg),