+++ pkg/cloud/cloud_api.go
@@ -4,7 +4,7 @@ package cloud
 import (
 	"context"
 
-	"github.com/newrelic/newrelic-client-go/pkg/errors"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/errors"
 )
 
 // Create or modify a cloud integration.
@@ -2588,7 +2588,6 @@ const getLinkedAccountsQuery = `query(
 			id
 			isEnabled
 			name
-			slug
 			updatedAt
 		}
 		slug