+++ pkg/cloud/types.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/newrelic/newrelic-client-go/pkg/nrtime"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/nrtime"
 )
 
 // CloudMetricCollectionMode - How metrics will be collected.